+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [2.0.0]
+
+### Added
+* TouchDiver Pro status message handling
+* Annular, pinky and palm actuation points
+* TouchDiver Pro tracking data message handling
+* TouchDiver Pro tracking raw data message handling
+
+### Changed
+* MiddlewareStatusUpdate now contains connectedDevices, devices (TouchDiver) and G2Devices (TouchDiver Pro)
+* Updated documentation and changelog
  
 ## [1.3.0]